+++ src/main/java/org/ga4gh/drs/utils/objectloader/FileDrsObjectLoader.java
@@ -78,7 +78,7 @@ public class FileDrsObjectLoader extends AbstractDrsObjectLoader {
         return checksums;
     }
 
-    public int imputeSize() {
+    public long imputeSize() {
         return (int)getFile().length();
     }
